android: back out delayLoad because it interacts badly with other things
Notably the delay is restarted fully on every load attempt, so we never
actually get to finish the actual load if we try to often.

This leads to content not being displayed if the view is refreshed often
(for example while there is a download running in the same listview).