interim test for group key generation by members. Fixed the XoClient disallowing simple group members to generate a new group key.
Currently there seems to still be an issue with what actually constitutes a finished invite.
In live situations this is probably not an issue since the users will react a lot slower that the tests can cascade the calls.

However this issue warrants more investigation as the borders of the state of a properly invited client do not seem to be entirely clear (to me at least).