Enable OrmLite object caching in integration test clients
Since object caching is enabled in the Android client, enabling it in
the integration tests as well makes them more similar to production.

It is important to not just enable the object cache, but to provide a
cache instance for every database backend. Otherwise, a global cache will
be used that confuses objects from different clients, leading to strange,
indeterministic test failures.