Fix flaky integration test
In this case the flakiness was caused by a call to Thread.sleep() with an
insufficient time parameter. Instead of using sleep, we identified a condition
that must be fulfilled before we can proceed, namely that an invited client
needs to receive a group key before it can join a group.

Hopefully fixes ITGroupJoin.joinGroupTestWhileAdminIsOffline.

Fixes part C of #464