Move intent check to onViewCreated in SingleProfileFragment
This fixes #474. Previously onStart got called again after selecting an Avatar
image which always called startActionMode() in onStart(). But in case the action
mode was already running it was cancelled internally which finished the whole
activity. Putting the intent check in onViewCreated ensures that this part is called
only once in the fragments lifetime.