Fix "multiple TalkClientContact references overwriting the database entries" correctly
As it happens the TalkClientContact class contains a ForeignCollection field which
is null on creation. This is checked by ormlite on create() which leads to the fact
that the instance is not put into the database cache(com.j256.ormlite.stmt.mapped.MappedCreate.java:116).
Therefore retrieving the contact from the database creates (and caches) a new
instance of the same contact data.
The fix just returns a newly retrieved instance instead of the created one.
Since TalkClientContact is the only class containing a ForeignCollection this only
happens with them.