Stop marking "deleted" contacts
Depending on the contact type, deleting means de-pairing, leaving or deleting
a group. Depending on the context, "deleted" contacts should be shown (e.g. in
group member lists) or not (e.g. in the contact list). Usages of "deleted" are
either redundant or can be replaced by queries for relationship, membership
and/or group state.