Fix missing admin in group member list
This is a workaround for a conceptual bug in OrmLite. We use the OrmLite
object cache (and rely on it for MediaCollections to work correctly). However,
OrmLite will also put "incomplete" (i.e. incorrect) instances, queried using
selectColumns(), into the object cache. As a result, we sometimes get wrong
objects from the cache when the same object has been queried before using a
selectColumns() clause. The workaround is not using selectColumns() at all.