Added initial support for worldwide support in xo client
The android part does not yet know anything about worldwide.
In fact the client does not even know properly if nearby is currently active without falling back to weird things like 'mEnvironmentGroupId'. This is just a member variable - ideally the XoClient would be in some kind of state regarding environments.