Workaround for not being able to bring many talktool clients up and running
It seems that the reason for this workaround is that bouncycastle is not threadsafe - at some level some locking is required probably. It seems that this locking on the client side (the only party actually doing crytpo) could help.
Especially SRP6VerifyingClient.calculateSecret() seems to have some form of issue.