The maxnumber of groups seems slightly off
Proposed something, that i still need to double check.

In my tests i had the case of having 2 groups, one group with only one client remaining and another group with 6 clients (7 clients in total). It seems to me that the merging should already have happened earlier.

The calculations are int/int divisions which is a remainder-less division.